--- src/components/circle2.js
@@ -1,4 +1,0 @@
-import createEntity from './entity2';
-import CircleEvents from '../events/circle';
-
-export default createEntity('Circle', 'center', CircleEvents);
