+++ src/components/entity.js
@@ -1,65 +1,85 @@
 import React from 'react';
-import createReactClass from 'create-react-class';
-import Listener from '../mixins/listener';
+// import createReactClass from 'create-react-class';
+// import Listener from '../mixins/listener';
 import compareProps from '../utils/compare-props';
 
-// export default createEntity('InfoWindow', 'position', InfoWindowEvents);
+// refactored to React class, then to React +16.3
 
 export default (name, latLngProp, events) => {
-  return createReactClass({
+	return class CreateEntity extends React.Component {
+		state = {
+			entity: null,
+		};
 
-    mixins: [Listener],
+		componentDidMount() {
+			const options = this.getOptions(this.props);
+			console.log('options', options);
+			this.setState({ entity: new google.maps[name](options) });
+			console.log('this.state.entity', this.state.entity);
+			this.addListeners(this.state.entity, events);
+		}
 
-    entity: null,
+		componentWillReceiveProps(nextProps) {
+			if (!compareProps(this.props, nextProps)) {
+				const options = this.getOptions(nextProps);
+				this.state.entity.setOptions(options);
+			}
+		}
 
-    componentDidMount() {
-      const options = this.getOptions(this.props);
-      console.log('options', options);
-      this.entity = new google.maps[name](options);
-      console.log('this.entity', this.entity);
-      this.addListeners(this.entity, events);
-    },
+		componentWillUnmount() {
+			this.state.entity.setMap(null);
+			this.removeListeners();
+			this.setState({ entity: null });
+		}
 
-    componentWillReceiveProps(nextProps) {
-      if (!compareProps(this.props, nextProps)) {
-        const options = this.getOptions(nextProps);
-        this.entity.setOptions(options);
-      }
-    },
+		getOptions(props) {
+			return {
+				...props,
+				[latLngProp]: this.switchPaths(name, props),
+			};
+		}
 
-    componentWillUnmount() {
-      this.entity.setMap(null);
-      this.removeListeners();
-      this.entity = null;
-    },
+		addListeners(entity, events) {
+			for (let prop in this.props) {
+				if (this.props.hasOwnProperty(prop) && events[prop]) {
+					const addListener = google.maps.event.addListener;
+					const listener = addListener(entity, events[prop], this.props[prop]);
+					if (!this.listeners) {
+						this.listeners = [];
+					}
+					this.listeners.push(listener);
+				}
+			}
+		}
 
-    getEntity() {
-      return this.entity;
-    },
+		getEntity() {
+			return this.state.entity;
+		}
 
-    getOptions(props) {
-      return {
-        ...props,
-        [latLngProp]: this.switchPaths(name, props)
-      };
-    },
+		switchPaths(name, props) {
+			switch (name) {
+				case 'Polyline':
+					return props.path;
+					break;
+				case 'Polygon':
+					return props.paths;
+					break;
+				default:
+					return new google.maps.LatLng(props.lat, props.lng);
+					break;
+			}
+		}
 
-    switchPaths(name, props) {
-      switch (name) {
-        case 'Polyline':
-          return props.path;
-          break;
-        case 'Polygon':
-          return props.paths;
-          break;
-        default:
-          return new google.maps.LatLng(props.lat, props.lng);
-          break;
-      }
-    },
-  
-    render() {
-      return null;
-    }
-  });
+		removeListeners() {
+			if (window.google && this.listeners) {
+				this.listeners.forEach(listener => {
+					google.maps.event.removeListener(listener);
+				});
+			}
+		}
+
+		render() {
+			return null;
+		}
+	};
 };
