--- src/components/info-window2.js
@@ -1,4 +1,0 @@
-import createEntity from './entity2';
-import InfoWindowEvents from '../events/info-window';
-
-export default createEntity('InfoWindow', 'position', InfoWindowEvents);
