--- src/components/marker2.js
@@ -1,4 +1,0 @@
-import createEntity from './entity2';
-import MarkerEvents from '../events/marker';
-
-export default createEntity('Marker', 'position', MarkerEvents);
