--- src/components/polygon2.js
@@ -1,4 +1,0 @@
-import createEntity from './entity2';
-import PolygonEvents from '../events/polygon';
-
-export default createEntity('Polygon', 'paths', PolygonEvents);
