--- src/components/polyline2.js
@@ -1,4 +1,0 @@
-import createEntity from './entity2';
-import PolylineEvents from '../events/polyline';
-
-export default createEntity('Polyline', 'path', PolylineEvents);
