+++ src/index.js
@@ -1,9 +1,9 @@
 import Gmaps from './components/gmaps';
-import Marker from './components/marker2';
-import InfoWindow from './components/info-window2';
-import Circle from './components/circle2';
-import Polyline from './components/polyline2';
-import Polygon from './components/polygon2';
+import Marker from './components/marker';
+import InfoWindow from './components/info-window';
+import Circle from './components/circle';
+import Polyline from './components/polyline';
+import Polygon from './components/polygon';
 
 export {Gmaps, Marker, InfoWindow, Circle, Polyline, Polygon};
